+++ spec/models/risk_spec.rb
@@ -6,6 +6,8 @@ RSpec.describe Risk, type: :model do
 
   it { is_expected.to belong_to(:escort) }
 
+  it { is_expected.to delegate_method(:editable?).to(:escort) }
+
   subject { described_class.new }
 
   include_examples 'reviewable'